requirements.txt
matplotlib==3.4.3
more_itertools==8.10.0
numpy==1.20.3
pandas==1.3.4
PyArabic==0.6.15
PyYAML==6.0
scikit_learn==0.24.2
scikit_learn==1.1.2
seaborn==0.11.2
torch==1.12.1+cu116
transformers==4.26.0
Unidecode==1.2.0
